MineRL/requirements.txt
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
cloudpickle==1.6.0
coloredlogs==15.0
cycler==0.10.0
decorator==5.0.6
dill==0.3.3
future==0.18.2
getch==1.0
gym==0.18.0
humanfriendly==9.1
idna==2.10
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
kiwisolver==1.3.1
lxml==4.6.3
matplotlib==3.0.3
minerl==0.3.6
numpy==1.20.2
opencv-python==4.5.1.48
pandas==1.2.4
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.18
psutil==5.8.0
ptyprocess==0.7.0
pyglet==1.5.0
Pygments==2.8.1
pyparsing==2.4.7
Pyro4==4.80
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
scipy==1.6.2
serpent==1.30.2
six==1.15.0
tqdm==4.60.0
traitlets==5.0.5
typing==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
